Separate post editor in preparation for being able to edit posts
